+++ Client/Client.Qt/Widgets/IPopupMessage.hpp
